+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unpublished]
 
+### Added
+- Loader for the simple OFF (Object File Format) format 
+
 ### Changed
 - Extended the in-memory structure